pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "spharagusnet"
version = "0.2.0"
description = "Neural document OCR for Chilean territorial planning documents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "J. Sebastian Miño"},
]
keywords = ["ocr", "neural-network", "chile", "urban-planning", "document-extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Text Processing :: General",
]

dependencies = [
    "Pillow>=10.0.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "torch>=2.2.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
train = [
    "pdfplumber>=0.11.0",
    "PyPDF2>=3.0.0",
    "python-Levenshtein>=0.25.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
]
dev = [
    "spharagusnet[train]",
    "build>=1.0.0",
    "twine>=5.0.0",
]

[project.scripts]
spharagusnet = "spharagusnet.__main__:main"

[project.urls]
Homepage = "https://github.com/jsmdev/spharagusnet"
Repository = "https://github.com/jsmdev/spharagusnet"
Issues = "https://github.com/jsmdev/spharagusnet/issues"

[tool.setuptools.packages.find]
include = ["spharagusnet*"]

[tool.setuptools.package-data]
spharagusnet = ["py.typed"]
